chore!: Remove timeout property on ArcjetRateLimitRule (#182)
This removes the `timeout` property on `ArcjetRateLimitRule`. Having timeout on rate limits is confusing because each rate limit has a reset time based on the configuration. If we kept timeout ("block duration") around, it would be very confusing to determine when a user will be able to access the resource again.

We're tracking the concept of timeout ("block duration") in https://github.com/arcjet/arcjet/issues/567.