feat!: Add fixedWindow, tokenBucket, and slidingWindow primitives (#184)
Replaces #171 

This adds the `fixedWindow`, `tokenBucket` and `slidingWindow` primitives. It keeps the `rateLimit()` primitive for backwards compatibility, but it is breaking because `protectSignup` was altered to require sliding window configurations.

The changes were streamlined by putting `requested` into the `extra` field of the RequestDetails. This allowed us to keep the same structure for specifying our Rules.

Draft because I need to add tests to maintain full test coverage and because it relies on #179, #180, #181, #182, and #183 (after which, I'll rebase all the cherry-picked commits out of this PR).