feat: Allow user-defined characteristics on rate limit options (#203)
Closes #202 
Closes https://github.com/arcjet/arcjet/issues/597

This adds TypeScript support for custom characteristics. While it was supported via the protocol, there was no indication to the user that they needed to add the field to every request.

By changing the types of our primitives, we can make the prop required on the request object (while filtering our well-known characteristics).

Draft because it needs some tests.