feat: Support cookies and query via the protocol (#214)
Closes #212 

This adds support for `cookies` and `query` fields on the protocol.

Previously, we only worked with cookies on the headers, but for ease-of-use, I moved them to a separate `cookies` field.
And we used to include the querystring as part of the path, but I moved it to a separate `query` field. This should make the path displays a lot nicer.

I also noticed we were always passing an empty string for the protocol, so I made some changes to that since we're parsing the URL for the querystring.