chore!: Switch Next.js to peer dependency (#339)
Closes #336

Next.js uses symbols on their objects, which breaks TypeScript compilation if you have version mismatches. This moves it to a peer dependency so the application version is always used. The peer dep version range is anything from 13, 14, and up because I don't really like putting an upper cap on it (what if Next.js 15 releases and it doesn't break our usage... it shouldn't cause churn for our users).