chore(example): Add SvelteKit app (#738)
I've gotten this working, finally. This time I looked into the `withRules` and used that in this implementations.

I still need to finish the "how it's done" part of the README, but I'd love a review and feedback on the code and install guide, make sure I'm not missing anything.