feat: Filter cookie headers when normalizing with ArcjetHeaders (#773)
Closes #215

This removes the `cookie` header when we normalize request headers to `ArcjetHeaders`. In #214, we moved the cookies into their own field on the `ArcjetRequest` so we no longer need it inside the `headers` field—sometimes cookies are very long so we should deduplicate that data.

The `ArcjetHeaders` constructor uses `this.append` throughout, so it made sense to put the filter in the `append` and `set` methods.