feat: Create Bun.sh adapter (#757)
This creates an Arcjet SDK adapter for [bun.sh](https://bun.sh/). This was much the same as the Node.js adapter, but I needed to introduce a new API to the adapter: `aj.handler()`. This API is used to wrap the request handler (called `fetch()` in Bun) to do some preprocessing on the request—in this case looking up the IP with `server.requestIP(request)` before storing it in a WeakMap.

Users *could* use this without using the `aj.handler()` function, but they'd need to have a proper IP existing in the headers, such as the `Fly-Client-IP` on fly.io; however, those won't exist when running locally and the `protect()` call will error without using `aj.handler()`.

TODO:
- [x] Add to CI and dependabot
- [x] SDK Stack in protobuf
- [x] Examples

Closes #475