fix(analyze): Leverage string interpolation to import Wasm files on edge runtime (#784)
While working on #775, I found that Vite doesn't support the `?module` syntax that we use for loading Wasm files in the edge runtime. Instead, we can use the `/* @vite-ignore */` comment but it only applies to imports using string interpolation.

I checked the output of `next build` and it seems to prepare all the `.wasm` files in our `./wasm/` directory as "edge chunks" when I use this syntax, so I think we should be good in both scenarios.