fix(analyze): Disable cache during base64 decode (#838)
Closes #836 

We are using `fetch` to base64 decode our Analyze Wasm files in certain runtimes. However, when this is used in the Next.js App Router, it fails because the URL is larger than 2mb. We don't actually need these to go through the `fetch` cache since it isn't making a network call, so we just disable it everywhere.