feat!: Allow ArcjetContext extension via new argument to core `protect()` (#841)
This is a breaking change to the core `arcjet` package because it adds a new `ctx` argument as the first parameter to `protect()`. This argument is meant to be used by adapters to provide additional information when calling `protect()`—for example, the `ARCJET_KEY` secret passed into Cloudflare's handler.  Most users won't notice this because they are using an adapter which don't expose the `ctx` argument to the user.