feat: Create sprintf package to replace util.format (#856)
This adds a new `@arcjet/sprintf` package to this monorepo, which provides a platform-independent formatter function. This will allow us to move closer to #51 by replacing the `util.format` import!

I've kept the implementation pretty close to the [reference library](https://github.com/pinojs/quick-format-unescaped) because [pino](https://github.com/pinojs/pino) uses it for formatting strings and we want pino to be a drop-in replacement for our logger when users want to [swap in structured logging](https://github.com/arcjet/arcjet-js/issues/855).