chore(logger)!: Replace nodejs util import with our sprintf library (#857)
Depends upon #856 

Towards #51 

This replaces the `import { format } from "util"` with our custom sprintf formatter. This is a breaking change to the logging library because we are more restrictive on data types; however, this doesn't really matter because we're making larger breaking changes for #822 and #855 before we release the beta version.