chore(decorate)!: Use console to log instead of Arcjet logger (#865)
This switches our `@arcjet/decorate` package to use `console.log` after formatting a string with our new sprintf package. Since I'm removing the singleton logger in #858 and this is a standalone package, it doesn't make sense to require users to inject a logger into this package.

I believe libraries like Pino have a way to intercept `console.log` so this should be fine for most usage.