feat!: Create runtime package and remove from SDK (#871)
This introduces a new `@arcjet/runtime` package to the monorepo. It provides a single function that returns a [WinterCG](https://wintercg.org/) identifier for the runtime.

Since this available as a standalone package, I removed the `runtime` API from the SDKs, but the `runtime` property is also available on the `ArcjetContext`.

I finally had to give up trying to get full test coverage because Node & Jest make it very difficult to change `process.release.name`.