feat!: Move all environment lookup into separate package (#897)
This introduces a new `@arcjet/env` package that provides various functions which receive an "env" object (such as `Bun.env` or `process.env`) and look for various keys that we care about. These keys are part of the `Env` type exported by the package to have a centralized definition of the properties.

By separating this logic, we remove all Node.js or other runtime specific logic in our generic packages. The "env" object lookup is then done in each adapter.

Closes #51 
Closes #885 

TODO:
- [x] Tests for env package
- [x] Docs for env package