deps: Add turborepo as a root devDependency (#1377)
This adds turborepo as a devDependency at the root of our monorepo. This will allow us to build the packages in order, which fixes an issue @e-moran encountered with the redact packages.

I chose turborepo because we already depend on some vercel packages so adding another doesn't expand our supply chain dependence.

Also, I've aliased all the npm scripts in our root to the turbo commands to make sure nothing changes from a usability standpoint.

Unfortunately, turborepo doesn't have a Wasm build so we need to allow optional dependencies again.