feat: add library to perform redaction of sensitive information (#1358)
Adds the library to perform redaction on text using the Arcjet sensitive info primitive.

Redaction is performed by the `redact()` function, which takes a `string` and any configuration that you want to use. It returns the redacted text and an `unredact()` function as a `[redacted, unredact]` tuple. Calling `unredact()` with a string will replace the fake data with the original data.