fix(arcjet): Infer types when no detect function is specified (#1446)
While trying to improve the type inference on #1358, I found that we weren't providing the correct type inference on the `sensitiveInfo` rule when a user **does not** specify a `detect()` function. This is likely the most common scenario so we want to have proper type inference here.

To solve this, I made `Detect` very generic throughout the types and then used some techniques (commented inline) to figure out if the function was specified in the options.