deps(example): Update all example deps (#1537)
This cherry picks all open dependabot PRs. It seems that Next.js put a breaking change in their latest release that adds a "salt hash" inside the swc compiler which fails if all versions of the compiler are not the same in the repo.

This requires further investigation to see if we need to change our dependabot setup, but we don't want to block the release.