fix(arcjet): Ensure performance measurements are 1-to-1 and always captured (#1858)
I moved a bunch of code around and into `try...finally` blocks so our `log.time()` and `log.timeEnd()` calls are 1-to-1 and always called (using the `finally` blocks). This means that the fingerprint generation isn't considered for the `local` label anymore, but is still timed by its own label.

The diff is really bad here because of the the new indentation.