chore(protocol)!: Remove received_at and decision fields from Report (#1988)
This updates our protocol package to the latest protobuf definition which removes the `decision` from `ReportResponse` to send less data down the wire which also reduces the amount of data we need to deserialize.

This means that the decision `id` field in the report log will be undefined in older SDKs so I've marked this as breaking.

It also removed `received_at` because our usage was inconsistent. We may revisit this in the future.