fix!: Return ERROR decision when fingerprint cannot be generated (#1990)
This changes the SDK to return an `ERROR` decision if we cannot build the fingerprint based on the data provided to `protect()`.  This is a breaking change because we used to continue with an empty fingerprint, but that wasn't the correct behavior.

I've also regenerated the Analyze Wasm and we've removed the logging in an attempt to check performance/memory usage which we'll investigate with examples and users. Ref #1724

Closes #1801