fix(ip)!: Exit early if platform-specific headers are missing IP (#2021)
As discussed at https://github.com/arcjet/arcjet-js/pull/2018#discussion_r1811564677, this changes the `findIP` function so it exits early if the platform-specific headers don't have a global IP address. Generally, this should never be the case, but if it somehow happened we would want to return an empty string instead of falling back to a different header that could be spoofed.