feat: Create nosecone package for creating secure headers (#2237)
This implements our `nosecone` package and 2 adapters, Next.js and SvelteKit. These 2 frameworks have some of the best support for nonce-based CSPv3—although Next.js has the caveat of it only working in dynamic mode.

Runtimes like Bun, Deno, and Node.js can use Nosecone directly to set headers on the responses, while adapters are needed for deeper integration. Using middleware works really well for Next.js because we can force the headers to be forwarded and it even detects the nonce from the `script-src` directive, which it adds to each `<script>` tag that webpack generates. For SvelteKit, we need to provide `csp` in the config so it'll add the CSP header, but we also use a hook to add our additional secure headers.

Notably missing:
- Removing X-Powered-By—the frameworks don't give us access to this, we should instead just recommend it be removed.
- A Remix adapter—it needs a lot of logic to thread nonces throughout the application.
- An Express adapter—most people are already using Helmet.