chore(protocol): Opt out of Buf & ConnectRPC v2 changes (#2473)
Buf v2 introduced support for "Protobuf Editions" and started embedding the protobuf definition as a base64 string in the output. Without trying to actually do the upgrade, this seems like a non-starter for serverless environments such as Vercel's Edge Runtime.

I updated our generator plugins to the latest version before v2 and restricted dependabot to avoid the upgrades.

We've had the general plan of a custom protocol and custom transports, but limiting upgrades may push us to doing it sooner.