chore: Switch most test harnesses to node:test (#2479)
This migrates most of our test harnesses to `node:test`. This will eventually allow us to remove Jest completely.

I still need to figure out how to make custom environments to test the `@arcjet/runtime` package and run tests with the [edge-runtime](https://edge-runtime.vercel.app/packages/runtime) package but this is a step in the right direction.