feat: Implement Astro integration (#2992)
This implements an Arcjet adapter for Astro. We do this via an integration that generates a virtual module, such that `arcjet()` gets added to the `integrations` array in `astro.config.mjs` with the normal SDK configuration. The integration then generates a client and types as the `arcjet:client` virtual module. This ensures there is a singleton client in the Astro module graph.

Closes #1075