deps(dev): Override esbuild to ensure newest version (#3205)
This overrides esbuild throughout the repo to ensure we are on the newest version. It is only in our dependency tree due to devDeps on frameworks and they all use `^` versioning which treats `0.24` vs `0.25` as incompatible, so we need to apply this as an override until it is resolved transitively.