fix(transport): Add connect as a direct dependency to satisfy peerDep (#3416)
Closes #3415 

It looks like whichever package manager is being used by the end-user is nesting the ConnectRPC dependencies which means that that `@connectrpc/connect` dependency included by the `@arcjet/protocol` package. This embeds the peerDependency as a direct dependency to ensure it is available even when package managers are being weird.