feat: Add `@arcjet/inspect` package with utilities to inspect decisions (#3455)
I've created a new package called `@arcjet/inspect` that provides utilities for inspecting decisions and the rule results provided on them.

The initial utilities are `isSpoofedBot`, `isVerifiedBot`, and `isMissingUserAgent` which were chosen because we've seen users encounter problems with them. Previously, we've documented the verbose way to handle these cases, but we'll be able to streamline to show these utilities.

It's worth noting that I've made all these utilities ignore `DRY_RUN` results because users wouldn't want dry run rules to be used in informing more detailed decisions. If users are trying to log information, they should iterate the results and use the type-narrowing functions attached to each result.