deps(example): Upgrade SvelteKit to eslint 9 & restrict Remix example (#3557)
This migrates the SvelteKit example to eslint 9. I generated a new project and then migrated over the new files and dependencies.

It seems that Remix doesn't [support](https://github.com/remix-run/remix/issues/10109) eslint 9, as none of their templates are updated. I'm guessing that it won't get these updates since it is soft-deprecated in favor of React Router v7.

Towards #539