fix(arcjet-node): Wrap `process.env` access with a getter object (#3559)
Closes #2986

This adds an abstraction over `process.env` in the Arcjet Node adapter. This is necessary because Gatsby uses Webpack with the replacer plugin to replace `process.env.SOMEVAR` with the string literal of the environment variable. However, we still want to support environment modification in proper Node.js runtime environments.

We may want to apply this same technique to all adapters but it is fragile so I erred on the side of caution.