feat: Support CIDR strings as proxies (#3577)
This adds CIDR parsing and matching to our `@arcjet/ip` package. I've also update all adapters to consume the `parseProxy` function when instantiating the SDK. If a CIDR matcher is encountered during the proxy comparison, the CIDR range is checked to see if it contains the IP segments.

I've chosen to implement this way because non-CIDR addresses remain the fastest, since it is just a string comparison and we want to pre-parse the CIDR addresses upon startup to ensure the fastest IP lookup we can achieve while still supporting CIDR proxies.

I've also tested all the Cloudflare ranges that they publish and I used https://ipinfo.io/tools/cidr-to-ip-range-converter to find the last address. Another utility that helped me explore CIDR was https://cidr.xyz/

Closes #2402