fix: Ensure local rules return DRY_RUN state in dry run mode (#3820)
This changes local rule results to set the state to `DRY_RUN` when the rule is configured in `DRY_RUN` mode. This ensures the rule results are returned the same locally and remotely. We'll also want to be able to set the state to `CACHED` in the future, which is how I discovered this. 

Fixes #3818