fix(runtime): Avoid warning during Edge Runtime compilation (#3905)
This (mostly) duplicates the `index.ts` file in our runtime package but removes the `process` identifier in the `edge-light` runtime. This is to avoid a warning from the Next.js compiler that warns about the `process` identifier even though it would never be used.

You can test this by setting up the `examples/nextjs-clerk-shield` app and running `npm run build` in it. Without the change, we see the warning, but after the change it is gone. You should also be able to check the CI for that example.

Closes #3876