fix(arcjet-node): Ensure `process.env.RENDER` is surfaced (#3969)
This fixes the follow up issue found in #3899, which showed up because of the giant `WARNING` I placed in the code and then promptly forgot about. I've made a change that leverages a TypeScript `satisfies` to ensure all keys are defined (try removing `RENDER` and notice the build failure). This also allowed me to remove the warning since it'll be caught by the compiler.

Closes #3899