fix(protocol): Ensure sensitiveInfo rule sends mode (#4026)
This fixes the `sensitiveInfo` rule which wasn't sending the `mode` field when converted to the Protobuf.

I've also refactored the way types were being specified to avoid casting rules. I found this while adding the `version` field but wanted to add it to this PR because it ensures the `ArcjetSensitiveInfoRule` is constructed correctly.