feat(protocol): Support identifier and version on rules (#4027)
This lays the groundwork of support for Rule IDs in the SDK. It updates the protobuf definitions and threads `version` and `rule_id` through the SDK.

It also implements the logic for stable, deterministic, and unique identifiers for local rules. These are mostly unused right now but are useful for double-checking the local and remote Rule IDs are exactly the same. 

These locally generated IDs will be needed for looking up cached decisions in the new cache implementation. Rules that don't currently generate an ID will get them added once they have a local implementation that reads from cache.