feat(arcjet)!: Require every rule to have a validate & protect function (#4204)
This removes the concept of "local" rules that are a superset of `ArcjetRule` by requiring all rules to have a `validate` and `protect` function when defined in the SDK.

I've marked this as breaking but it would only be breaking if someone wrote a custom rule that didn't have a local component, which doesn't really make sense because it wouldn't execute anything on the server.

This sets the foundation for #4191 to lookup cache values in every rule. 

Context: In chatting with @qw-in, we realized that it made sense to join the concept of "local" and "remote" rules in the SDK to just `ArcjetRule`. Even if they don't have any implementation locally, every rule will leverage the `protect()` function to look up values in the cache. This also simplifies the rule concept a bit when we work with `@arcjet/protocol` in the SDK.