chore(arcjet): Reduce timeouts and other transport issues to info log (#4411)
This changes the log level of an inactionable error from our transport from being logged in users' error logs. I noticed that SDK timeout errors were showing up in our system but they are inactionable. This indicates to me that this should be at a lower log level. I chose `INFO` because that is the same level that used by the Report RPC call to log any transport errors.