test: clean protocol tests (#4479)
In GH-4415 and GH-4472 I noticed that in particular the protocol tests could use some cleaning up.
I had some extra time to work on a big PR due to holiday in US.

Test cases were often testing equality of big objects as a sort of “fixture” testing. This made it difficult to see *what* was different between test cases. And, if a property would be added to such objects, all tests would need changing.

After cleaning (the first 2 commits), I went through the uncovered lines and added tests cases for them.

Importantly, there are barely any changes to `client.ts` and `index.ts`. Only because their behavior was not tested, and not allowed by the types, I removed the dead code.

***Please open the files that GH collapses, and tip: go through commits one by one***