refactor: sort packages in release please config (#4589)
In the release video by Blaine, Quinn noted that `@arcjet/cache` appeared to be missing.
After some searching, Blaine commented that it was there, just that the sorting was not up to date.
This solves that: makes sure everything is sorted again, potentially preventing a future problem where a newly added package was forgotten.

I was wondering whether the sorting may be important: e.g., `arcjet` was pretty high before all the integrations, making it publish before integrations, which may seem like a good idea. However, `arcjet` *depends* on packages further down. Defeating that purpose. So as that argument doesn’t hold, it can just as well be sorted alphabetically.

With sorting, I do put `analyze-wasm` *before* `analyze`. So sorting `-` before EOF. The inverse can also be argued. I can see arguments for either way. And myself slightly prefer this.

An adjacent idea that the reviewer will probably note, that can be discussed now, but can also be deferred to a later point, is that some folders use an `arcjet-` prefix whereas others don’t. Specifically, integrations. So, `env` refers to `@arcjet/env` and `arcjet-bun` refers to `@arcjet/bun`. I don‘t like the inconsistency, perhaps we can do without. I see that having the integrations next to each other is nice, but now this block breaks up the unprefixed utilities. Perhaps just folders such as `integrations/bun/` is an alternative way to group integrations.