docs: add info on derivate works to license sections
Several readmes have `## Implementation` sections.
Most but not all of those sections relate to which derivate
works the implementation was inspired by and how it differs.

While this information often doesn’t have to be top of mind,
it is important for people who wonder about license stuff.
So I think it is best to put it *inside* the license section.
Using a similar format.