docs: add uniform use section to readmes
This PR adds a uniform use section to everything.

Many packages had few examples.
Several didn’t work anymore.
Packages that integrate with other tools have many examples on the website.
`@arcjet/next` had a lot of (outdated) information.

This adds one `## Use` section right under `## Install` that shows a *minimal*
example of how a package can work.
My idea is that you can paste it into an `example.ts`,
install things,
and it works.
It’s not a goal to show everything that’s possible with a package.

For that reason,
I opted to only show `shield` with the integrations.
And to have them all do the same things,
so that people can see how Arcjet is (typically) the same and only the other
tools differ.

How to use rules in your client,
and all other ways to use arcjet and integrations,
is then left to the website.

Note that I do want to add `## API` sections too.
I *do* want to cover everything that’s exposed in readmes.
Just not in the top-most simple example.

Related-to: GH-4338.