test: clean `ip` tests (#4632)
In the IP tests, there were functions making functions making objects making assertions.
It was difficult to wrap my head around what was going on.
Functions have some benefits but also makes it hard to keep track of what’s going on, creating arguably potentially smelly code.
The alternative is to be a bit more verbose in each test case. Less functions to build things. Loops and arrays for repetition.
I did that.

Some examples of arguable smells were: 

* exact same tests running many times (https://redirect.github.com/arcjet/arcjet-js/blob/03c5da93b33728a1d9b561ef72b4a7e0bf010fcc/ip/test/ipv4.test.ts#L9, note that `make` is not used in these)
* some tests don’t exactly fit the abstractions, so then you need to attach them separately and do things manually (https://redirect.github.com/arcjet/arcjet-js/blob/03c5da93b33728a1d9b561ef72b4a7e0bf010fcc/ip/test/ipv6.test.ts#L205)
* similarly, options need to be passed to at different abstractions, the `platform` option deals with which headers are supported but the `proxy` option which IP is supported, but then the question is, what option takes precedence at what abstraction? What is really tested?
* a split down ipv4 and ipv6, with similar copy/pasted code

That being said, I importantly did not *change* the tests. Didn’t add, move, or remove. But there is room for that.
I do wonder whether we really need to test every potential valid/invalid IP (64 tests) at every header or other property in the request.
More important seems a split around platforms, to test that most headers are *not* supported if a platform is set. Which is documented in the code but currently not tested.

Still, for the same number of lines, I do think it’s now more obvious what is going on.
I left the commits so work progress can be seen, but that shows a lot of the in-between-state, I’d recommend side by side.