build(examples): simplify reusable examples (#4692)
This removes all the repetition per example folder and instead uses a matrix for all the repetition. Cutting down 1.2k lines!

There is a teeny tiny downside that Astro requires an `ARCJET_KEY` and the ±4 Clerk examples require a `NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY`, which are now set for the build of every example. But, these are both fake tokens. So no danger. So not a real problem in my opinion.

I did look for a second at the missing examples. They are missing for various reasons. Which can be solved in a follow up:

build broken:
- `nextjs-react-hook-form`

build missing:
- `express-sensitive-info`
- `express-bots`
- `express-newman`
- `nodejs-express-rate-limit`
- `nodejs-express-launchdarkly`
- `nodejs-express-validate-email`
- `nodejs-rate-limit`

bun:
- `bun-hono-rate-limit`
- `bun-rate-limit`

deno:
- `deno-sensitive-info`

Related-to: GH-4666 😈