test: add tests for unsupported headers per platform (#4694)
I noticed that the implementation specifically does not support most headers if a particular `platform` is set while working on GH-4632. And, in some cases, when a platform is not set, several platform-specific headers are not supported either. Behavior that makes sense when you think about it.

This commit thus adds checks for when headers are not supported.

Related-to: GH-4632.