refactor(examples): move `characteristics` to rules
To solve GH-4553, I looked for `characteristics`.
I moved every case to a rule, when that type-checked.
If there was no rule used that accepted them,
I assumed that they were not needed,
and removed them.

I only found top-level use inside the examples,
so I think this closes GH-4553.