refactor(arcjet): use `interface`s to share fields
To prepare for more JSDocs in `arcjet` (core),
I switch to interfaces.
Interfaces support a nice chain of extension which
helps for our system of mutually exclusive options.

However, `interface`s are in some small ways more strict
than putting some object into a `type`:
in certain cases they do not support indexing into them.
This is a good thing but it introduces a type issue so
I added some code to handle non-objects as options.