refactor(redact): reuse exposed types
Some of the types from `@arcjet/redact-wasm` are not used anywhere.
Otherwise were redefined in `@arcjet/redact` itself.
This commit removes the duplication and removes the (completely
internal) unused API surface from `@arcjet/redact-wasm`.

Some of the exposed API surface did not specify a return type either.
This adds return types so that we can verify that the API matches what
we want instead of returning arbitrary objects.