refactor(sprintf): replace abbreviations with words
This refactors the internal parameters used in the printer with
full words, based a bit on Pino:
<https://github.com/pinojs/pino/blob/main/docs/api.md#trace>.
This is useful because we document these parameters and abbreviations
are jargon not always understood by everyone.
Compared to Pino, this is a smaller utility doing interpolation,
so I went with `values` instead of a longer `interpolationValues`
or so.
And, in some cases, the constant `length` was stored in a variable
(`valuesLength`).
That felt a bit too much.

Related-to: GH-4853.