refactor(ip): rename identifiers to match other casing (#4723)
When working on the fastify integration, i used `@arcjet/ip`.

I noticed that I needed the `Cidr` type, which is a return value of `parseProxy` from there.
So I expose it here.

I also noticed that there are 2 functions, `parseProxy` as a named export and `findIp` as the default export.
I expose the latter as a named export too, and deprecate the default symbol.

Internally, the casing of abbreviations was more capital heavy, and not always consistent (`parseCIDR` and `isIPv4Cidr`). Especially around symbols such as `isIPv4Cidr` (previous casing), I find it hard to come up with good arguments why that `P` is uppercase but `v` lowercase.
I like using capitals only for word boundaries: `parseCidr`, `isIpv4Cidr`, `Cidr`.
So, before exposing values, I changed their internal names. See GH-4478 for more on that.