chore(decorate): add type errors for invalid values
The `decorate` function accepts objects in 3 different shapes.
Code for that exists to check which shape the given value has.
However, a sort of manually combined shape was accepted as an
option.
This meant that people could pass objects that looked completely
different in.
That looked nothing like those 3 shapes.
This removes the superfluous code and introduces a type error
if someone passed an unknown value in.