refactor(redact): improve types around (non)standard entities
Previously, there were 2 type parameters that contained first an entire detect
function and a *separate* type parameter that dealt with which entities were
searched for (in `entities`), detected in `detect`, and replaced in `replace`.
That second type parameter was a sort of `string`, a little bit smarter.
Though, in several cases TypeScript used just a regular `string`.

This commit changes to a single type parameter of the exact `entities` being
searched for, detected, and replaced.
As that’s the value that needs to be smart and everything else is based on.
It extends from the known standard tags and also allows custom tags with the
TypeScript trick `(string & {})`.

These improved types help.
There is now good autocomplete when writing options.
And a bug was caught in the tests, where a value was replaced in a `replace`
function, but that value was not searched for in `entities`.